In progress Feature #14602 Create API client
https://redmine.biodiversity.agr.gc.ca/issues/14602

-Removed Operations component, replaced it with a context-provided
 doOperations function.
-Removed Operations component docs.
